Fix: Correctly capture openstack port cli output

+++ shell/openstack-cleanup-orphaned-ports.sh
@@ -36,23 +36,28 @@ mapfile -t os_ports_ts < <(openstack --os-cloud "$os_cloud" port list \
         -f value \
         -c ID \
         -c status \
-        -c created_at \
-        | grep -E DOWN \
-        | awk -F' ' '{print $1 " " $3}')
+        | grep -E DOWN )
 
 if [ ${#os_ports_ts[@]} -eq 0 ]; then
     echo "No orphaned ports found."
 else
     cutoff=$(date -d "30 minutes ago"  +%s)
     for port_ts in "${os_ports_ts[@]}"; do
-        created_at_isots="${port_ts#* }"
         port_uuid="${port_ts% *}"
-        echo "checking port uuid: ${port_uuid} with TS: ${created_at_isots}"
-        created_at_uxts=$(date -d "${created_at_isots}" +"%s")
-        # Clean up ports where created_at > 30 minutes
-        if [[ "$created_at_uxts" -gt "$cutoff" ]]; then
-            echo "Removing orphaned port $port_uuid created_at ts > 30 minutes."
-            openstack --os-cloud "$os_cloud" port delete "$port_uuid"
+        echo "Retrieving timestamp for port uuid: ${port_uuid}"
+        created_at=`openstack port show fc8789b7-06dc-40a8-8739-f06d472896d3 -f value -c created_at`
+        created_at_uxts=$(date -d "${created_at}" +"%s")
+        # Validate timestamp is numeric value
+        if [[ "$created_at_uxts" -eq "$created_at_uxts" ]]; then
+                # Clean up ports where created_at > 30 minutes
+                if [[ "$created_at_uxts" -gt "$cutoff" ]]; then
+                        echo "Removing orphaned port $port_uuid created_at ts > 30 minutes."
+                        openstack --os-cloud "$os_cloud" port delete "$port_uuid"
+                else
+                        echo "Port did not meet requirement for deletion; moving on"
+                fi
+        else
+                echo "Date variable failed numeric test; deletion not possible"
         fi
     done
 fi
